requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
decorator==5.1.0
Deprecated==1.2.13
django-scheduler==0.9.5
graphviz==0.19.1
icalendar==4.0.9
idna==3.3
importlib-metadata==4.10.0
jieba==0.42.1
numpy==1.21.5
objgraph==3.5.0
packaging==21.3
paddlepaddle-tiny==1.6.1
protobuf==3.19.1
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.0
requests==2.26.0
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
tornado==6.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wrapt==1.13.3
zipp==3.6.0
